refactor(gateway): extract the review runner from review.ts
Moves runOne, drainOnShutdown, backoff, replyWithTs and the inFlight
registry into ReviewRunner. These are the only members that touch
inFlight, so the shared state moves with them rather than staying spread
across a coordinator that no longer uses it.

The acquire -> work -> release bracket moved WHOLE, not split: its
protocol (every early exit passes through the finally, which reads
`posted` to decide the claim's fate) stays locally verifiable.

review.ts 909 -> 548, under the 800 ceiling. Public API unchanged;
slack/index.ts and every test untouched.
